my alarm is now working. wth

+++ executable.c
@@ -15,7 +15,7 @@
 typedef struct Clock
 {
         int sec;
-        int nsec;
+        long nsec;
 } Clock;
 
 struct msgbuf
